+++ internal/pkg/instrumentation/bpf/net/http/server/probe.go
@@ -95,8 +95,11 @@ func New(logger logr.Logger) probe.Probe {
 				Val: structfield.NewID("std", "net/http", "Request", "Proto"),
 			},
 		},
-		Uprobes: map[string]probe.UprobeFunc[bpfObjects]{
-			"net/http.serverHandler.ServeHTTP": uprobeServeHTTP,
+		Uprobes: []probe.Uprobe[bpfObjects]{
+			{
+				Sym: "net/http.serverHandler.ServeHTTP",
+				Fn:  uprobeServeHTTP,
+			},
 		},
 
 		ReaderFn: func(obj bpfObjects) (*perf.Reader, error) {