+++ internal/pkg/instrumentation/manager.go
@@ -24,6 +24,7 @@ import (
 
 	dbSql "go.opentelemetry.io/auto/internal/pkg/instrumentation/bpf/database/sql"
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/bpf/github.com/gin-gonic/gin"
+	otelTraceGlobal "go.opentelemetry.io/auto/internal/pkg/instrumentation/bpf/go.opentelemetry.io/otel/traceglobal"
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/bpf/google.golang.org/grpc"
 	grpcServer "go.opentelemetry.io/auto/internal/pkg/instrumentation/bpf/google.golang.org/grpc/server"
 	httpClient "go.opentelemetry.io/auto/internal/pkg/instrumentation/bpf/net/http/client"
@@ -34,9 +35,6 @@ import (
 	"go.opentelemetry.io/auto/internal/pkg/process"
 )
 
-// Error message returned when unable to find all instrumentation functions.
-var errNotAllFuncsFound = fmt.Errorf("not all functions found for instrumentation")
-
 // Manager handles the management of [probe.Probe] instances.
 type Manager struct {
 	logger         logr.Logger
@@ -44,10 +42,11 @@ type Manager struct {
 	done           chan bool
 	incomingEvents chan *probe.Event
 	otelController *opentelemetry.Controller
+	globalImpl     bool
 }
 
 // NewManager returns a new [Manager].
-func NewManager(logger logr.Logger, otelController *opentelemetry.Controller) (*Manager, error) {
+func NewManager(logger logr.Logger, otelController *opentelemetry.Controller, globalImpl bool) (*Manager, error) {
 	logger = logger.WithName("Manager")
 	m := &Manager{
 		logger:         logger,
@@ -55,6 +54,7 @@ func NewManager(logger logr.Logger, otelController *opentelemetry.Controller) (*
 		done:           make(chan bool, 1),
 		incomingEvents: make(chan *probe.Event),
 		otelController: otelController,
+		globalImpl:     globalImpl,
 	}
 
 	err := m.registerProbes()
@@ -103,10 +103,8 @@ func (m *Manager) FilterUnusedProbes(target *process.TargetDetails) {
 			}
 		}
 
-		if n := len(inst.Manifest().Symbols); funcsFound != n {
-			if funcsFound > 0 {
-				m.logger.Error(errNotAllFuncsFound, "some of expected functions not found - check instrumented functions", "instrumentation_name", name, "funcs_found", funcsFound, "funcs_expected", n)
-			}
+		if funcsFound == 0 {
+			m.logger.Info("no functions found for probe, removing", "name", name)
 			delete(m.probes, name)
 		}
 	}
@@ -209,6 +207,10 @@ func (m *Manager) registerProbes() error {
 		httpClient.New(m.logger),
 		gin.New(m.logger),
 		dbSql.New(m.logger),
+	}
+
+	if m.globalImpl {
+		insts = append(insts, otelTraceGlobal.New(m.logger))
 	}
 
 	for _, i := range insts {